moves process info environment variables (client ID and license key) to AppSecret.swift
